swtpm: Use Encrypt-then-MAC when writing encrypted data
Version 1 of the encryption was prepending the hash on the plaintext
data then decrypting all of it, which is not so good. In version 2
we now use Encrypt-then-MAC (EtM) where we encrypt the data and then
calculate an HMAC on the encrypted data and then preprend the HMAC.
We bump up the version in the header from 1 to 2 and remove the code
writing the hash in front of it but keep the part that was reading it.

Unfortunately there is the data that's written to files where we do
not have the version information as we have it for blobs exported
via the API. Here we have to try the decryption following version 2
first, then try version 1.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>